# river: reuse memory when decoding a map