--- a/pkg/river/internal/value/decode.go
+++ b/pkg/river/internal/value/decode.go
@@ -389,11 +389,16 @@ func decodeObject(val Value, rt reflect.Value) error {
 		// Create a shared value to decode each element into. This will be zeroed
 		// out for each key, and then copied when setting the map index.
 		into := reflect.New(rt.Type().Elem()).Elem()
+		intoZero := reflect.Zero(into.Type())
 
-		for _, key := range val.Keys() {
+		for i, key := range val.Keys() {
 			// We ignore the ok value because we know it exists.
 			value, _ := val.Key(key)
 
+			// Zero out the value if it was decoded in the previous loop.
+			if i > 0 {
+				into.Set(intoZero)
+			}
 			// Decode into our element.
 			if err := decode(value, into); err != nil {
 				return FieldError{Value: val, Field: key, Inner: err}
@@ -401,9 +406,6 @@ func decodeObject(val Value, rt reflect.Value) error {
 
 			// Then set the map index.
 			res.SetMapIndex(reflect.ValueOf(key), into)
-
-			// Zero out the value for the next iteration.
-			into.Set(reflect.Zero(into.Type()))
 		}
 
 		rt.Set(res)