Benchmark results:

```
name             old time/op    new time/op    delta
ObjectDecode-10    5.83µs ± 2%    5.68µs ± 2%   -2.54%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
ObjectDecode-10    3.97kB ± 0%    3.67kB ± 0%   -7.66%  (p=0.000 n=9+10)

name             old allocs/op  new allocs/op  delta
ObjectDecode-10       125 ± 0%       106 ± 0%  -15.20%  (p=0.000 n=10+10)
```

This gives a tiny performance boost for decoding map values, but it creates far less garbage.

Other low-hanging fruit optimizations are possible, but I wanted to start with this one.